requirements.txt
matplotlib
numpy
opencv-python
Pillow
scikit-image
tensorboard
torch
torchaudio
torchvision
